// src/shared/enums/area.enum.ts
export enum AreaEnum {
  HOME = 'HOME',
  MUN = 'MUN',
  ESC = 'ESC',
  PRO = 'PRO',
  ALU = 'ALU',
  TRF_ALU = 'TRF_ALU',
  ENT_ALU = 'ENT_ALU',
  LAN_RES = 'LAN_RES',
  INF = 'INF',
  MEN = 'MEN',
  REL = 'REL',
  LANC = 'LANC',
  ENTU = 'ENTU',
  SINT_GER = 'SINT_GER',
  DESC = 'DESC',
  NIV_DES = 'NIV_DES',
  LIN_EVO = 'LIN_EVO',
  EVO_LEI = 'EVO_LEI',
  AVA = 'AVA',
  AVA_ON = 'AVA_ON',
  EDI = 'EDI',
  EDI_MUN = 'EDI_MUN',
  TES = 'TES',
  MAT_REF = 'MAT_REF',
  JOR_PED = 'JOR_PED',
  ANO_LET = 'ANO_LET',
  DISC = 'DISC',
  SER = 'SER',
  TUR = 'TUR',
  USU = 'USU',
  PER_ACES = 'PER_ACES',
  IMP_EXP = 'IMP_EXP',
  MCO = 'MCO',
  LOG_SIS = 'LOG_SIS',
  EXP_DAD = 'EXP_DAD',
  REL_EXT = 'REL_EXT',
  REL_IFR = 'REL_IFR',
  REL_NAL = 'REL_NAL',
  REL_COR = 'REL_COR',
  SIN_TEST = 'SIN_TEST',
  REG_UNI = 'REG_UNI',
  REG_EST = 'REG_EST',
  REG_MUN = 'REG_MUN',
  HIST_DES = 'HIST_DES',
  LST_RES = 'LST_RES',
  MEN_TUT = 'MEN_TUT',
  AI_ASSIST = 'AI_ASSIST',
  OPR_LOTE = 'OPR_LOTE',
}